requirements.txt
Pillow
numpy
send2trash
